Rename camera directory to camera-manager.
