+++ src/components/timeline.ts
@@ -2,7 +2,7 @@ import { CSSResultGroup, html, LitElement, TemplateResult, unsafeCSS } from 'lit
 import { customElement, property } from 'lit/decorators.js';
 import timelineStyle from '../scss/timeline.scss';
 import { CameraConfig, ExtendedHomeAssistant, TimelineConfig } from '../types';
-import { CameraManager } from '../camera/manager';
+import { CameraManager } from '../camera-manager/manager';
 import { View } from '../view/view';
 import './surround.js';
 import './timeline-core.js';
